added get req on front end and back end, renders markers from database on page load

+++ server/database/models/markers.js
@@ -5,8 +5,8 @@ const Markers = db.define('markers', {
   _id: {type: DataTypes.INTEGER, allowNull: false, primaryKey: true, autoIncrement: true},
   commonName: {type: DataTypes.STRING, allowNull: false, foreignKey: false},
   time: {type: DataTypes.DATE, allowNull: false, foreignKey: false},
-  lat: {type: DataTypes.INTEGER, allowNull: false, foreignKey: false},
-  lng: {type: DataTypes.INTEGER, allowNull: false, foreignKey: false},
+  lat: {type: DataTypes.DECIMAL(10,4), allowNull: false, foreignKey: false},
+  lng: {type: DataTypes.DECIMAL(10,4), allowNull: false, foreignKey: false},
  });
 
  module.exports = {
